Performance and coding improvements Some coding improvements like intializing lists with the correct length instead of growing them one by one, and some better handling if if and else statements.
In the across method, I hard coded some options in the first `matchit` calls because we don't need those set; the point of that call is only to estimate PS, not discard any units or perform any matching. `method = "nearest"` is the fastest.

Also changed how the resulting dataset is created. It now uses the same structure as weightthem, i.e., just collectign the weights and other components from the matchit output and attaching them to the existing dataset. Anyone who was not matched automatically gets a weight of 0. This dramatically speeds up the function.